[refactor][frontend] Atualizando as rotas de importação para o Mentor.ts
